requirements.txt
streamlit==1.38.0
langchain-groq==0.2.0
python-dotenv==1.0.0
streamlit-authenticator==0.3.2
pyyaml==6.0.2
requests==2.32.3
mcp==0.1.0
pytest==8.3.3
pytest-mock==3.14.0